move operations into own module
